+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # FFXIV Style Durability Bars
+### 1.2.0
+* Updated .tocs for Classic interface version 11508, TBC interface version 20505, MOP interface version 50503, Midnight interface version 120000
+* Additional options for removing bars for items without durability, and rarity colours for bars
+* Durability percentages in tooltips
 ### 1.1.3
 * Updated .tocs for Classic interface version 11507, MOP interface version 50500, TWW interface version 110107
 ### 1.1.2